[issue414] cleanup incomplete plan files in get_plan_files()

+++ src/driver/portfolio_runner.py
@@ -31,6 +31,8 @@ import traceback
 
 DEFAULT_TIMEOUT = 1800
 
+PLAN_COST_REGEX = re.compile(r"; cost = (\d+) \((unit cost|general cost)\)\n")
+
 # Exit codes.
 EXIT_PLAN_FOUND = 0
 EXIT_CRITICAL_ERROR = 1
@@ -67,49 +69,61 @@ def set_limit(kind, soft, hard):
               (kind, (soft, hard), err, resource.getrlimit(kind)), file=sys.stderr)
 
 
+def plan_file_is_complete(plan_file):
+    with open(plan_file) as input_file:
+        return any(PLAN_COST_REGEX.match(line) for line in input_file)
+
+
 def get_plan_cost_and_cost_type(plan_file):
     with open(plan_file) as input_file:
         for line in input_file:
-            match = re.match(r"; cost = (\d+) \((unit cost|general cost)\)\n", line)
+            match = PLAN_COST_REGEX.match(line)
             if match:
                 return int(match.group(1)), match.group(2)
-    os.remove(plan_file)
-    print("Could not retrieve plan cost from %s. Deleted the file." % plan_file)
-    return None, None
+    assert False
 
 
 def get_plan_file(plan_prefix, number):
     return "%s.%d" % (plan_prefix, number)
 
 
-def get_plan_files(plan_prefix):
+def get_plan_files_and_cleanup(plan_prefix):
+    """Return all complete plan files.
+
+    If the last plan file is incomplete, delete it.
+    """
     plan_files = []
     for index in itertools.count(start=1):
         plan_file = get_plan_file(plan_prefix, index)
         if os.path.exists(plan_file):
-            plan_files.append(plan_file)
+            if plan_file_is_complete(plan_file):
+                plan_files.append(plan_file)
+            else:
+                os.remove(plan_file)
+                print("%s is incomplete. Deleted the file." % plan_file)
+                break
         else:
             break
     return plan_files
 
 
 def get_cost_type(plan_prefix):
-    for plan_file in get_plan_files(plan_prefix):
-        _, cost_type = get_plan_cost_and_cost_type(plan_file)
-        if cost_type is not None:
-            return cost_type
+    """This method is only called after a plan has been found."""
+    plan_files = get_plan_files_and_cleanup(plan_prefix)
+    assert plan_files
+    _, cost_type = get_plan_cost_and_cost_type(plan_files[0])
+    return cost_type
 
 
 def get_g_bound_and_number_of_plans(plan_prefix):
     plan_costs = []
-    for plan_file in get_plan_files(plan_prefix):
+    for plan_file in get_plan_files_and_cleanup(plan_prefix):
         plan_cost, _ = get_plan_cost_and_cost_type(plan_file)
-        if plan_cost is not None:
-            if plan_costs and not plan_costs[-1] > plan_cost:
-                raise SystemExit(
-                    "Plan costs must decrease: %s" %
-                    " -> ".join(str(c) for c in plan_costs + [plan_cost]))
-            plan_costs.append(plan_cost)
+        if plan_costs and not plan_costs[-1] > plan_cost:
+            raise SystemExit(
+                "Plan costs must decrease: %s" %
+                " -> ".join(str(c) for c in plan_costs + [plan_cost]))
+        plan_costs.append(plan_cost)
     bound = min(plan_costs) if plan_costs else "infinity"
     return bound, len(plan_costs)
